+++ Backend/Project/Models/Subscription.cs
@@ -6,7 +6,7 @@ public class Subscription
 {
     public int Id { get; set; }
 
-    public double DefaultPrice { get; set; }
+    public decimal DefaultPrice { get; set; }
     
     [NotMapped]
     public int DurationInDays
